chapter-1-lab/requirements.txt
streamlit
langchain
langchain_google_genai
python-dotenv

